Redesigned forgot password dialog
