+++ .idea/misc.xml
@@ -16,6 +16,7 @@
         <entry key="..\:/Dev/Work/android/pass-vault-kotlin-android/app/src/main/res/layout/dialog_change_passcode.xml" value="0.33" />
         <entry key="..\:/Dev/Work/android/pass-vault-kotlin-android/app/src/main/res/layout/dialog_credits.xml" value="0.33" />
         <entry key="..\:/Dev/Work/android/pass-vault-kotlin-android/app/src/main/res/layout/dialog_forgot_password.xml" value="0.67" />
+        <entry key="..\:/Dev/Work/android/pass-vault-kotlin-android/app/src/main/res/layout/dialog_verify_secret.xml" value="0.3625" />
         <entry key="..\:/Dev/Work/android/pass-vault-kotlin-android/app/src/main/res/layout/fragment_create_login.xml" value="0.33" />
         <entry key="..\:/Dev/Work/android/pass-vault-kotlin-android/app/src/main/res/layout/fragment_generate_password.xml" value="0.75" />
         <entry key="..\:/Dev/Work/android/pass-vault-kotlin-android/app/src/main/res/layout/fragment_login.xml" value="0.1" />
