+++ app/src/androidTest/java/com/vinsonb/password/manager/kotlin/fragments/LoginFragmentTest.kt
@@ -11,8 +11,6 @@ import androidx.preference.PreferenceManager
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.vinsonb.password.manager.kotlin.R
@@ -72,14 +70,5 @@ class LoginFragmentTest {
         }
 
         Assert.assertEquals(navController.currentDestination?.id, R.id.view_accounts_fragment)
-    }
-
-    @Test
-    fun forgotPasswordButton_clicked_showsDialogVerifySecret() {
-        onView(withId(R.id.button_forgot_password))
-            .perform(click())
-
-        onView(withId(R.id.layout_dialog_verify_secret_answer))
-            .check(matches(isDisplayed()))
     }
 }
