+++ app/src/main/res/layout/dialog_forgot_password.xml
@@ -4,19 +4,41 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:padding="8dp">
+    android:padding="16dp">
+
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/layout_dialog_secret_answer"
+        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:hint="What is your mother's maiden name?">
+
+        <com.google.android.material.textfield.TextInputEditText
+            android:id="@+id/input_dialog_secret_answer"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:inputType="text"
+            android:textColorHint="#757575"
+            tools:ignore="DuplicateSpeakableTextCheck,TextContrastCheck"
+            tools:text="Doe" />
+
+    </com.google.android.material.textfield.TextInputLayout>
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/layout_dialog_change_passcode"
         style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:hint="@string/hint_passcode"
+        android:layout_marginTop="16dp"
+        android:hint="@string/hint_new_passcode"
         app:endIconDrawable="@drawable/ic_see_passcode"
         app:endIconMode="password_toggle"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
+        app:layout_constraintTop_toBottomOf="@id/layout_dialog_secret_answer">
 
         <com.google.android.material.textfield.TextInputEditText
             android:id="@+id/input_dialog_change_passcode"
@@ -36,7 +58,7 @@
         android:layout_width="0dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="16dp"
-        android:hint="@string/hint_repeat_passcode"
+        android:hint="@string/hint_new_repeat_passcode"
         app:endIconDrawable="@drawable/ic_see_passcode"
         app:endIconMode="password_toggle"
         app:layout_constraintEnd_toEndOf="parent"
